framework for script
