+++ package.json
@@ -19,7 +19,6 @@
     "lint": "tsc --noEmit && eslint src/**/*.ts{,x}",
     "lint:fix": "tsc --noEmit && eslint src/**/*.ts{,x} --fix",
     "nodemon": "nodemon",
-    "whiteboard-checkedout-no-user": "ts-node-dev src/scripts/whiteboard-checkedout-no-user.ts",
     "delete-space-force": "ts-node-dev src/delete-space-force.ts",
     "delete-organization": "ts-node-dev src/delete-organization.ts",
     "remove-user-from-organizations": "ts-node-dev src/remove-user-from-organizations.ts",
@@ -34,6 +33,7 @@
     "migration-add-default-discussion-callouts": "ts-node-dev src/migration/add-default-discussion-callouts.ts",
     "update-application-forms-on-space-nl": "ts-node-dev src/scripts/application-form-dutch.ts",
     "convert-challenge-callout": "ts-node-dev src/scripts/convert-challenge-callout.ts",
+    "spaces-license-usage-excel": "ts-node-dev src/spaces-license-usage-excel/spaces-license-usage-excel.ts",
     "prepublishOnly": "npm run build",
     "codegen": "graphql-codegen --config codegen.yml"
   },
